requirements.txt
asgiref==3.8.1
crispy-bootstrap5==0.7
Django==4.2.7
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-decouple==3.8
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
requests==2.31.0
